changed speed type to uint16_t from float

+++ Core/Inc/util/CMSIS_extra/global_variables_handler.h
@@ -16,7 +16,7 @@ typedef struct {
     uint16_t wheels[4];
 } SPEEDS_t;
 typedef struct {
-    float wheels[WHEEL_ENCODERS_AVAILABLE];
+    uint16_t wheels[WHEEL_ENCODERS_AVAILABLE];
 } WHEEL_ENCODER_SPEEDS_t;
 typedef uint16_t FRONT_AVG_SPEED_t;
 typedef uint16_t REAR_AVG_SPEED_t;
